test: deepen and harden the mitata benchmark suite
- Split benchmark/index.js into theme files (throughput / abort /
  nested) behind a thin index.js entry that parses an optional name
  filter and a --json flag.
- Harden: .gc('inner') on the allocation-heavy groups (overhead,
  bufferSize, input shape, nested) to remove cross-iteration GC noise;
  run({ throw: true }) so a broken bench fails the process loudly.
- Deepen: new "abort & error delivery" group (pre-aborted signal,
  mid-stream external abort, fail-fast trigger, fail-eventually
  AggregateError accumulation) and a sync-iterable / array input
  comparison exercising makeIterableAsync — all previously
  unbenchmarked code paths.
- Add fixtures: syncRange, rejectingCallback.
- Add `npm run bench:json` for capture/diff against a local baseline.
- CLAUDE.md: refresh the Benchmarks section for the new layout.
- README.md: add a "## Performance" section with the stable
  qualitative findings and relative ratios.
