+++ CLAUDE.md
@@ -86,6 +86,12 @@ For testing rejections, prefer the `.catch(err => ({ rejectedWith: err }))` enve
 
 ## Benchmarks
 
-`npm run bench` runs the [mitata](https://github.com/evanwashere/mitata) suite in `benchmark/` (`node --expose-gc --allow-natives-syntax`). One `group()` per major design decision: library-overhead tax, always-on abort wiring cost, ordered vs unordered dispatch, `bufferSize` scaling, nested sub-iterators, and the `mergeIterables` wrapper. They exist to catch performance regressions when the state machine changes — re-run before/after any change to `fillQueue`, `nextValue`, or the abort wiring.
+`npm run bench` runs the [mitata](https://github.com/evanwashere/mitata) suite (`node --expose-gc --allow-natives-syntax`). `benchmark/fixtures.js` holds the shared helpers; three theme files register benches on import and `benchmark/index.js` is the only one that calls `run()`:
 
-The non-negotiable rule, enforced by `benchmark/fixtures.js`: **benchmark fixtures never use timers.** `asyncRange` yields on the microtask queue with no artificial delay, so the numbers reflect the library's per-item bookkeeping overhead and not simulated I/O. A `setTimeout`-based source would make every benchmark measure the event loop instead. `do_not_optimize` wraps the drained result so the JIT cannot eliminate the loop; mitata handles warmup and flags suspicious (dead-code-eliminated) results with a `!`. Numbers are only comparable within a single run on the same machine — there is no committed baseline to diff against; trust the *relative* `summary()` output, not absolute times.
+- `benchmark/throughput.js` — overhead vs raw `for await`, `bufferSize` scaling, ordered vs unordered dispatch, input shape (async generator / sync iterable / array).
+- `benchmark/abort.js` — always-on abort-wiring cost, plus abort & error *delivery* (pre-aborted signal, mid-stream external abort, `fail-fast` triggering, `fail-eventually` `AggregateError` aggregation).
+- `benchmark/nested.js` — nested sub-iterators, `mergeIterables` wrapper overhead.
+
+`npm run bench:json` emits JSON — capture it before and after a change for a local diff (no baseline is committed; the numbers are machine-specific). `node benchmark/index.js <regex>` passes a filter to `run()`, matched against **bench names** (so `… abort` runs every bench with "abort" in its name, not a whole group). `run({ throw: true })` makes a broken bench fail the process loudly.
+
+The non-negotiable rule, enforced by `benchmark/fixtures.js`: **benchmark fixtures never use timers.** `asyncRange` / `syncRange` yield with no artificial delay, so the numbers reflect the library's per-item bookkeeping overhead and not simulated I/O. A `setTimeout`-based source would make every benchmark measure the event loop instead. `do_not_optimize` wraps the drained result so the JIT cannot eliminate the loop; mitata handles warmup and flags dead-code-eliminated results with a `!`. The allocation-heavy groups (overhead, `bufferSize`, input shape, nested) use `.gc('inner')` — GC before each iteration — to remove cross-iteration GC noise; this makes those benches slower but their distribution tighter. The abort & error *delivery* benches are deliberately left on the default `.gc('once')`: they are composite metrics (construct + partial consume + teardown) and run noisier than the steady-state throughput benches — read them for the *shape* of the teardown cost, not a precise number. Re-run before/after any change to `fillQueue`, `nextValue`, or the abort wiring.
