+++ package.json
@@ -30,6 +30,7 @@
   ],
   "scripts": {
     "bench": "node --expose-gc --allow-natives-syntax benchmark/index.js",
+    "bench:json": "node --expose-gc --allow-natives-syntax benchmark/index.js --json",
     "build": "run-s build:*",
     "build:0": "run-s clean",
     "build:1-declaration": "tsc -p declaration.tsconfig.json",
